% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_comma_space.R
\name{add_comma_space}
\alias{add_comma_space}
\title{Ensure Space After Comma}
\usage{
add_comma_space(x)
}
\arguments{
\item{text.var}{The text variable.}
}
\value{
Returns a vector of strings with commas that have a space after them.
}
\description{
Adds a space after a comma as \code{strip} and many other functions may consider a 
comma separated string as one word (i.e., \code{"one,two,three"} becomes 
\code{"onetwothree"}  rather than \code{"one two three"}).
}
\examples{
\dontrun{
x <- c("the,  dog,went", "I,like,it", "where are you", NA, "why", ",", ",f")
add_comma_space(x)
}
}
\keyword{comma}
\keyword{space}

