% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing_endmark.R
\name{add_missing_endmark}
\alias{add_missing_endmark}
\title{Add Missing Endmarks}
\usage{
add_missing_endmark(x, replacement = "|", endmarks = c("?", ".", "!"), ...)
}
\arguments{
\item{x}{The text variable.}

\item{replacement}{Character string equal in length to pattern or of length 
one which are  a replacement for matched pattern.}

\item{endmarks}{The potential ending punctuation marks.}

\item{\dots}{Additional arguments passed to 
\code{\link[textclean]{has_endmark}}.}
}
\value{
Returns a vector with missing endmarks added.
}
\description{
Detect missing endmarks and replace with the desired symbol.
}
\examples{
x <- c(
    "This in a", 
    "I am funny!", 
    "An ending of sorts\%", 
    "What do you want?"
)

add_missing_endmark(x)
}
