% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmod-package.R
\docType{data}
\name{DATA}
\alias{DATA}
\title{Fictitious Classroom Dialogue}
\format{A data frame with 11 rows and 5 variables}
\usage{
data(DATA)
}
\description{
A fictitious dataset useful for small demonstrations.
}
\details{
\itemize{
  \item person. Speaker
  \item sex. Gender
  \item adult. Dummy coded adult (0-no; 1-yes)
  \item state. Statement (dialogue)
  \item code. Dialogue coding scheme
}
}
\keyword{datasets}

