% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmod-package.R
\docType{data}
\name{hamlet}
\alias{hamlet}
\title{Hamlet (Complete & Split by Sentence)}
\format{A data frame with 2007 rows and 7 variables}
\usage{
data(hamlet)
}
\description{
A dataset containing the complete dialogue of Hamlet with turns of talk split
into sentences.
}
\details{
\itemize{
  \item act. The act (akin to repeated measures)
  \item tot. The turn of talk
  \item scene. The scene (nested within an act)
  \item location. Location of the scene
  \item person. Character in the play
  \item died. Logical coded death variable if yes the character dies in the
  play
  \item dialogue. The spoken dialogue
}
}
\references{
\url{http://www.gutenberg.org}
}
\keyword{datasets}

