% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_endmark.R
\name{has_endmark}
\alias{has_endmark}
\title{Test for Incomplete Sentences}
\usage{
has_endmark(x, endmarks = c("?", ".", "!"), ...)
}
\arguments{
\item{x}{A character vector.}

\item{endmarks}{The potential ending punctuation marks}

\item{\dots}{ignored.}
}
\value{
Returns a logical vector.
}
\description{
A logical test of missing sentence ending punctuation.
}
\examples{
x <- c(
    "I like it.", 
    "Et tu?",  
    "Not so much", 
    "Oh, I understand.",  
    "At 3 p.m., we go",
    NA
)
has_endmark(x)
}
\keyword{incomplete}

