% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_text.R
\name{print.check_text}
\alias{print.check_text}
\title{Prints a check_text Object}
\usage{
\method{print}{check_text}(x, include.text = TRUE, file = NULL, ...)
}
\arguments{
\item{x}{The check_text object.}

\item{include.text}{logical.  If \code{TRUE} the offending text is printed as 
well.}

\item{file}{A connection, or a character string naming the file to print to.
If \code{NULL} prints to the console.}

\item{\ldots}{ignored}
}
\description{
Prints a check_text object.
}
