% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_text_logicals.R
\name{print.which_are_locs}
\alias{print.which_are_locs}
\title{Prints a which_are_locs Object}
\usage{
\method{print}{which_are_locs}(x, n = 100, file = NULL, ...)
}
\arguments{
\item{x}{A which_are_locs object}

\item{n}{The number of affected elements to print out (the rest are 
truncated)}

\item{file}{Path to an external file to print to}

\item{\ldots}{ignored.}
}
\description{
Prints a which_are_locs object
}
