% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_row.R
\name{remove_row}
\alias{remove_empty_row}
\alias{remove_row}
\title{Remove Rows That Contain Markers}
\usage{
remove_row(dataframe, search.column, terms, keep.rownames = FALSE, ...)

remove_empty_row(dataframe)
}
\arguments{
\item{dataframe}{A dataframe object.}

\item{search.column}{Column name to search for markers/terms.}

\item{terms}{The regex terms/markers of the rows that are to be removed from 
the dataframe.}

\item{\ldots}{Other arguments passed to \code{\link[base]{grepl}}.}
}
\value{
\code{remove_row} - returns a dataframe with the termed/markered rows 
removed.

\code{remove_empty_row} - returns a dataframe with empty rows removed.
}
\description{
\code{remove_row} - Remove rows from a data set that contain a given marker/term.

\code{remove_empty_row} - Removes the empty rows of a data set that are common in 
reading in data.
}
\examples{
\dontrun{
#remove_row EXAMPLE:
remove_row(DATA, "person", c("sam", "greg"))
remove_row(DATA, 1, c("sam", "greg"))
remove_row(DATA, "state", c("Comp"))
remove_row(DATA, "state", c("I "))
remove_row(DATA, "state", c("you"), ignore.case=TRUE)

#remove_empty_row EXAMPLE:
(dat <- rbind.data.frame(DATA[, c(1, 4)], matrix(rep(" ", 4), 
   ncol =2, dimnames=list(12:13, colnames(DATA)[c(1, 4)]))))
remove_empty_row(dat)
}
}

