% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_date.R
\name{replace_date}
\alias{replace_date}
\title{Replace Dates With Words}
\usage{
replace_date(x, pattern = NULL, replacement = NULL, ...)
}
\arguments{
\item{x}{The text variable.}

\item{pattern}{Character date regex string to be matched in the given 
character vector.}

\item{replacement}{A function to operate on the extracted matches or a 
character string which is a replacement for the matched pattern.}

\item{\ldots}{ignored.}
}
\value{
Returns a vector with the pattern replaced.
}
\description{
Replaces dates with word equivalents.
}
\examples{
x <- c(NA, '11-16-1980 and 11/16/1980', "and 2017-02-08 but then there's 2/8/2017 too")

replace_date(x)
replace_date(x, replacement = '<<DATE>>')
}
