% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_kerning.R
\name{replace_kerning}
\alias{replace_kerning}
\title{Replace Kerning Space with No Space}
\usage{
replace_kerning(x, ...)
}
\arguments{
\item{x}{The text variable.}

\item{\ldots}{ignored.}
}
\value{
Returns a vector with kern spaces removed.
}
\description{
In typography kerning is the adjustment of spacing.  Often in informal writing 
a form of kerning and all capital letters is used for emphasis.  This tool looks 
for3 or mor consecutive capital letters with spaces in between and removes the 
spaces.
}
\examples{
x <- c(
    "Welcome to A I: the best W O R L D!",
    "Hi I R is the B O M B for sure: we A G R E E indeed.",
    "A sort C A T indeed!",
    NA
)

replace_kerning(x)
}
\references{
\url{https://stackoverflow.com/a/47438305/1000343}
}
\author{
StackOverflow user @ctwheels
}
