% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_misspelling.R
\name{replace_misspelling}
\alias{replace_misspelling}
\title{Replace Misspelled Words}
\usage{
replace_misspelling(x, ...)
}
\arguments{
\item{x}{A character vector.}

\item{\ldots}{ignored..}
}
\value{
Returns a vector of strings with misspellings replaced.
}
\description{
Replace misspelled words with their most likely replacement.  This function 
uses \pkg{hunspell} in the backend.  \pkg{hunspell}  must be installed in 
order to use this feature.
}
\note{
The function splits the string apart into tokens for speed
optimization.  After the replacement occurs the strings are pasted back
together.  The strings are not guaranteed to retain exact spacing of the
original.
}
\examples{
bad_string <- c("I cant spelll rigtt noow.", '', NA, 
    'Thiss is aslo mispelled?', 'this is 6$ and 38 cents in back2back!')
replace_misspelling(bad_string)
}
\author{
Surin Space and Tyler Rinker <tyler.rinker@gmail.com>.
}
