% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_non_ascii.R
\name{replace_non_ascii}
\alias{replace_non_ascii}
\title{Replace Non-ASCII Characters}
\usage{
replace_non_ascii(x, replacement = "", ...)
}
\arguments{
\item{x}{The text variable.}

\item{replacement}{A string to replace incomplete punctuation marks with.}

\item{\dots}{ignored.}
}
\value{
Returns a text variable (character sting) with non-ascii characters 
replaced.
}
\description{
Replaces non-ascii characters.
}
\examples{
x <- c("Hello World", "6 Ekstr\\xf8m", "J\\xf6reskog", "bi\\xdfchen Z\\xfcrcher")
Encoding(x) <- "latin1"
x

replace_non_ascii(x)
replace_non_ascii(x, replacement="<<FLAG>>")
}
\keyword{ascii}

