% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_white.R
\name{replace_white}
\alias{replace_white}
\title{Remove Escaped Characters}
\usage{
replace_white(x, ...)
}
\arguments{
\item{x}{The character vector.}

\item{\dots}{ignored.}
}
\value{
Returns a vector of character strings with escaped characters removed.
}
\description{
Pre-process data to replace one or more white space character with a single 
space.
}
\examples{
x <- "I go \\r
    to   the \\tnext line"
x
replace_white(x)
}
\keyword{character}
\keyword{escaped}
