% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem.R
\name{stem_strings}
\alias{stem_strings}
\title{Stem a Vector of Strings}
\usage{
stem_strings(x, language = "porter", ...)
}
\arguments{
\item{x}{A vector of strings.}

\item{language}{The name of a recognized language (see 
\code{\link[SnowballC]{wordStem}}).}

\item{\ldots}{Other arguments passed to \code{\link[textshape]{split_token}}.}
}
\value{
Returns a vector of stemmed strings.
}
\description{
Stem a vector of strings.
}
\examples{
stem_strings(
    c(
        NA, 
        "The doggies, well they aren't joyfully running.", 
        "The daddies are coming over...", 
        "This is 34.546 above"
    )
)
}
\seealso{
\code{\link[textclean]{stem_words}}
}

