% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem.R
\name{stem_words}
\alias{stem_words}
\title{Stem a Vector of Words}
\usage{
stem_words(x, language = "porter", ...)
}
\arguments{
\item{x}{A vector of words.}

\item{language}{The name of a recognized language (see 
\code{\link[SnowballC]{wordStem}}).}

\item{\ldots}{ignored.}
}
\value{
Returns a vector of stemmed words.
}
\description{
Stem a vector of words.
}
\seealso{
\code{\link[textclean]{stem_strings}}
}

