% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_incomplete.R
\name{replace_incomplete}
\alias{replace_incomplete}
\title{Denote Incomplete End Marks With "|"}
\usage{
replace_incomplete(x, replacement = "|", ...)
}
\arguments{
\item{x}{The text variable.}

\item{replacement}{A string to replace incomplete punctuation marks with.}

\item{\dots}{ignored.}
}
\value{
Returns a text variable (character sting) with incomplete sentence 
marks (.., ..., .?, ..?, en & em dash etc.) replaced with "|".
}
\description{
Replaces incomplete sentence end marks (.., ..., .?, ..?, en & em dash etc.)
with \code{"|"}.
}
\examples{
x <- c("the...",  "I.?", "you.", "threw..", "we?")
replace_incomplete(x)
replace_incomplete(x, '...')
}
\keyword{incomplete-sentence}
